Split inclusion integrator from inclusion evolver
