+++ source/dynamics/inclusion_evolver.cpp
@@ -1,5 +1,5 @@
 /***************************************************************************
- *            dynamics/differential_inclusion.cpp
+ *            inclusion_evolver.cpp
  *
  *  Copyright  2008-20  Luca Geretti, Pieter Collins, Sanja Zivanovic
  *
@@ -30,12 +30,12 @@
 
 namespace Ariadne {
 
-TimeStepType lower_bound(TimeStepType const& t) {
-    return TimeStepType(FloatDPLowerBound(t,DoublePrecision()).raw());
-}
-
-TimeStepType lower_bound(Real const& t) {
-    return TimeStepType(FloatDPLowerBound(t,DoublePrecision()).raw());
+FloatDP volume(Vector<ApproximateIntervalType> const& box) {
+    FloatDP result = 1.0;
+    for (auto i: range(box.size())) {
+        result *= box[i].width().raw();
+    }
+    return result;
 }
 
 BoxDomainType initial_ranges_to_box(RealVariablesBox const& var_ranges) {
@@ -50,9 +50,9 @@ BoxDomainType initial_ranges_to_box(RealVariablesBox const& var_ranges) {
 struct ScheduledApproximator
 {
     Nat step;
-    InclusionApproximatorHandle approximator;
+    InclusionIntegratorHandle approximator;
 
-    ScheduledApproximator(Nat s, InclusionApproximatorHandle a) : step(s), approximator(a) {}
+    ScheduledApproximator(Nat s, InclusionIntegratorHandle a) : step(s), approximator(a) {}
 };
 
 inline OutputStream& operator<<(OutputStream& os, ScheduledApproximator const& sa) {
@@ -63,21 +63,20 @@ struct ScheduledApproximatorComparator {
         return (sa1.step > sa2.step); }
 };
 
-
 class InclusionEvolverState {
   private:
     Nat _step;
     List<ScheduledApproximator> _schedule;
-    Map<InclusionApproximatorHandle,Nat> _approximator_check_delay;
-    Map<InclusionApproximatorHandle,Nat> _approximator_global_optima_count;
-    Map<InclusionApproximatorHandle,Nat> _approximator_local_optima_count;
+    Map<InclusionIntegratorHandle,Nat> _approximator_check_delay;
+    Map<InclusionIntegratorHandle,Nat> _approximator_global_optima_count;
+    Map<InclusionIntegratorHandle,Nat> _approximator_local_optima_count;
   public:
     InclusionEvolverState(InclusionVectorField const& ivf, List<InputApproximation> const& approximations, IntegratorInterface const& integrator)
         : _step(0u)
     {
-        InclusionApproximatorFactory factory(integrator);
+        InclusionIntegratorFactory factory(integrator);
         for (auto appro : approximations) {
-            InclusionApproximatorHandle approximator = factory.create(ivf.function(),ivf.inputs(),appro);
+            InclusionIntegratorHandle approximator = factory.create(ivf.function(),ivf.inputs(),appro);
             _schedule.push_back(ScheduledApproximator(0u,approximator));
             _approximator_global_optima_count[approximator] = 0;
             _approximator_local_optima_count[approximator] = 0;
@@ -87,16 +86,16 @@ class InclusionEvolverState {
 
     Nat step() const { return _step; }
     List<ScheduledApproximator> const& schedule() const { return _schedule; }
-    Map<InclusionApproximatorHandle,Nat> const& check_delay() const { return _approximator_check_delay; }
-    Map<InclusionApproximatorHandle,Nat> const& global_optima_count() const { return _approximator_global_optima_count; }
-    Map<InclusionApproximatorHandle,Nat> const& local_optima_count() const { return _approximator_local_optima_count; }
+    Map<InclusionIntegratorHandle,Nat> const& check_delay() const { return _approximator_check_delay; }
+    Map<InclusionIntegratorHandle,Nat> const& global_optima_count() const { return _approximator_global_optima_count; }
+    Map<InclusionIntegratorHandle,Nat> const& local_optima_count() const { return _approximator_local_optima_count; }
 
     Void reset_local_optima_count() {
         for (auto entry : _approximator_local_optima_count)
             _approximator_local_optima_count[entry.first] = 0;
     }
 
-    Void update_with_best(InclusionApproximatorHandle const& best) {
+    Void update_with_best(InclusionIntegratorHandle const& best) {
         for (auto appro : approximators_to_use()) {
             if (best == appro) _reset_check_delay(appro);
             else _increase_check_delay(appro);
@@ -109,12 +108,12 @@ class InclusionEvolverState {
         _increase_optima_count(best);
     }
 
-    Void append_to_schedule(Nat const& step, InclusionApproximatorHandle const& approximator) {
+    Void append_to_schedule(Nat const& step, InclusionIntegratorHandle const& approximator) {
         _schedule.push_back(ScheduledApproximator(step,approximator));
     }
 
-    List<InclusionApproximatorHandle> approximators_to_use() const {
-        List<InclusionApproximatorHandle> result;
+    List<InclusionIntegratorHandle> approximators_to_use() const {
+        List<InclusionIntegratorHandle> result;
         for (auto i = _schedule.size(); i > 0; --i) {
             auto entry = _schedule[i-1];
             if (entry.step == _step) {
@@ -139,9 +138,9 @@ class InclusionEvolverState {
     }
 
   private:
-    Void _reset_check_delay(InclusionApproximatorHandle const& approximator) { _approximator_check_delay[approximator] = 0; }
-    Void _increase_check_delay(InclusionApproximatorHandle const& approximator) { _approximator_check_delay[approximator]++; }
-    Void _increase_optima_count(InclusionApproximatorHandle const& approximator) {
+    Void _reset_check_delay(InclusionIntegratorHandle const& approximator) { _approximator_check_delay[approximator] = 0; }
+    Void _increase_check_delay(InclusionIntegratorHandle const& approximator) { _approximator_check_delay[approximator]++; }
+    Void _increase_optima_count(InclusionIntegratorHandle const& approximator) {
         _approximator_global_optima_count[approximator]++;
         _approximator_local_optima_count[approximator]++;
     }
@@ -158,14 +157,14 @@ inline char activity_symbol(SizeType step) {
 }
 
 
-inline Map<InclusionApproximatorHandle,FloatDP> convert_to_percentages(Map<InclusionApproximatorHandle,Nat> const& approximation_global_frequencies) {
+inline Map<InclusionIntegratorHandle,FloatDP> convert_to_percentages(Map<InclusionIntegratorHandle,Nat> const& approximation_global_frequencies) {
 
     Nat total_steps(0);
     for (auto entry: approximation_global_frequencies) {
         total_steps += entry.second;
     }
 
-    Map<InclusionApproximatorHandle,FloatDP> result;
+    Map<InclusionIntegratorHandle,FloatDP> result;
     for (auto entry: approximation_global_frequencies) {
         result[entry.first] = 1.0/total_steps*entry.second;
     }
@@ -173,178 +172,7 @@ inline Map<InclusionApproximatorHandle,FloatDP> convert_to_percentages(Map<Inclu
     return result;
 }
 
-C1Norms::C1Norms(FloatDPError const& K_,Vector<FloatDPError> const& Kj_,FloatDPError const& pK_,Vector<FloatDPError> const& pKj_,
-             FloatDPError const& L_,Vector<FloatDPError> const& Lj_,FloatDPError const& pL_,Vector<FloatDPError> const& pLj_,
-             FloatDPError const& H_,Vector<FloatDPError> const& Hj_,FloatDPError const& pH_,Vector<FloatDPError> const& pHj_,
-             FloatDPError const& expLambda_,FloatDPError const& expL_)
- : K(K_), Kj(Kj_), pK(pK_), pKj(pKj_), L(L_), Lj(Lj_), pL(pL_), pLj(pLj_), H(H_), Hj(Hj_), pH(pH_), pHj(pHj_), expLambda(expLambda_), expL(expL_) {
-    _dimension = Kj.size();
-    assert(Kj.size() == _dimension and pKj.size() == _dimension and Lj.size() == _dimension and pKj.size() == _dimension and Hj.size() == _dimension && pHj.size() == _dimension);
-}
-
-Tuple<FloatDPError,Vector<FloatDPError>,FloatDPError,Vector<FloatDPError>,FloatDPError,Vector<FloatDPError>,FloatDPError,Vector<FloatDPError>,FloatDPError,Vector<FloatDPError>,FloatDPError,Vector<FloatDPError>,FloatDPError,FloatDPError>
-C1Norms::values() const {
-    return std::tie(this->K,this->Kj,this->pK,this->pKj,this->L,this->Lj,this->pL,this->pLj,this->H,this->Hj,this->pH,this->pHj,this->expLambda,this->expL);
-}
-
-
-C1Norms
-compute_norms(EffectiveVectorMultivariateFunction const& noise_independent_component, Vector<EffectiveVectorMultivariateFunction> const& input_derivatives, BoxDomainType const& inputs, PositiveFloatDPValue const& h, UpperBoxType const& B) {
-
-    auto n = noise_independent_component.result_size();
-    auto m = input_derivatives.size();
-    DoublePrecision pr;
-    FloatDPError ze(pr);
-    FloatDPError K=ze, pK=ze, L=ze, pL=ze, H=ze, pH=ze;
-    Vector<FloatDPError> Kj(n), pKj(n), Lj(n), pLj(n), Hj(n), pHj(n);
-    FloatDPUpperBound Lambda=ze;
-
-    auto Df=noise_independent_component.differential(cast_singleton(B),2);
-    for (auto j : range(n)) {
-        auto Df_j=Df[j].expansion();
-        FloatDPError K_j=ze, L_j=ze, H_j=ze; FloatDPUpperBound Lambda_j=ze;
-        for (auto ac : Df_j) {
-            UniformReference<MultiIndex> a=ac.index();
-            UniformReference<FloatDPBounds> c=ac.coefficient();
-            if (a.degree()==0) {
-                K_j += mag(c);
-            } else if (a.degree()==1) {
-                L_j += mag(c);
-                if (a[j]==1) { Lambda_j += c.upper(); }
-                else { Lambda_j += mag(c); }
-            } else {
-                assert(a.degree()==2);
-                H_j += mag(c);
-            }
-        }
-        K=max(K,K_j); L=max(L,L_j); H=max(H,H_j); Lambda=max(Lambda,Lambda_j);
-        Kj[j] = K_j;
-        Lj[j] = L_j;
-        Hj[j] = H_j;
-    }
-
-    Matrix<FloatDPError> pK_matrix(m,n), pL_matrix(m,n), pH_matrix(m,n);
-
-    for (auto i : range(m)) {
-        auto Dg_i=input_derivatives[i].differential(cast_singleton(B),2);
-        FloatDPError Vi(abs(inputs[i]).upper());
-        FloatDPError pK_i=ze, pL_i=ze, pH_i=ze;
-        for (auto j : range(n)) {
-            auto Dg_ij=Dg_i[j].expansion();
-            FloatDPError pK_ij=ze, pL_ij=ze, pH_ij=ze;
-            for (auto ac : Dg_ij) {
-                UniformReference<MultiIndex> a=ac.index();
-                FloatDPBounds const& c=ac.coefficient();
-                if (a.degree()==0) {
-                    pK_ij += mag(c);
-                } else if (a.degree()==1) {
-                    pL_ij += mag(c);
-                } else {
-                    assert(a.degree()==2);
-                    pH_ij += mag(c);
-                }
-            }
-            pK_i=max(pK_i,pK_ij); pL_i=max(pL_i,pL_ij); pH_i=max(pH_i,pH_ij);
-            pK_matrix[i][j] += pK_ij; pL_matrix[i][j] += pL_ij; pH_matrix[i][j] += pH_ij;
-        }
-
-        pK+=Vi*pK_i; pL+=Vi*pL_i; pH+=Vi*pH_i;
-    }
-
-    for (auto j : range(n)) {
-        pKj[j] = ze; pLj[j] = ze; pHj[j] = ze;
-        for (auto i : range(m)) {
-            FloatDPError Vi(abs(inputs[i]).upper());
-            pKj[j] += Vi*pK_matrix[i][j]; pLj[j] += Vi*pL_matrix[i][j]; pHj[j] += Vi*pH_matrix[i][j];
-        }
-    }
-
-    FloatDPError expLambda = (possibly(Lambda>0)) ? FloatDPError(dexp(Lambda*h)) : FloatDPError(1u,pr);
-    FloatDPError expL = cast_positive(exp(L*h));
-
-    return C1Norms(K,Kj,pK,pKj,L,Lj,pL,pLj,H,Hj,pH,pHj,expLambda,expL);
-}
-
-FloatDPError zeroparam_worstcase_error(C1Norms const& n, PositiveFloatDPValue const& h) {
-    return min(n.pK*n.expLambda*h,
-               (n.K*2u+n.pK)*h); }
-FloatDPError zeroparam_component_error(C1Norms const& n, PositiveFloatDPValue const& h, SizeType j) {
-    return min(n.pK*n.expL*h,
-               (n.Kj[j]*2u+n.pKj[j])*h); }
-
-FloatDPError oneparam_worstcase_error(C1Norms const& n, PositiveFloatDPValue const& h) {
-    return pow(h,2u)*((n.K+n.pK)*n.pL/3u +
-                      n.pK*2u*(n.L+n.pL)*n.expLambda); }
-FloatDPError oneparam_component_error(C1Norms const& n, PositiveFloatDPValue const& h, SizeType j) {
-    return n.pLj[j]*(n.K+n.pK)*pow(h,2u)/3u +
-           ((n.Lj[j]+n.pLj[j])*2u*n.pK)*cast_positive(cast_exact((n.L*n.expL*h+1u-n.expL)/pow(n.L,2u))); }
-
-template<> FloatDPError twoparam_worstcase_error<AffineInputs>(C1Norms const& n, PositiveFloatDPValue const& h, FloatDPError const& r) {
-    return ((r*r+1u)*n.pL*n.pK +
-            (r+1u)*h*n.pK*((n.pH*2u*r + n.H)*(n.K+r*n.pK)+pow(n.L,2u)+(n.L*3u*r+n.pL*r*r*2u)*n.pL)*n.expLambda +
-            (r+1u)/6u*h*(n.K+n.pK)*((n.H*n.pK+n.L*n.pL)*3u+(n.pH*n.K+n.L*n.pL)*4u)
-           )/cast_positive(+1u-h*n.L/2u-h*n.pL*r)*pow(h,2u)/4u; }
-template<> FloatDPError twoparam_worstcase_error<AdditiveInputs>(C1Norms const& n, PositiveFloatDPValue const& h, FloatDPError const& r) {
-    return (n.H*(n.K+n.pK)/2u +
-            (pow(n.L,2u)+n.H*(n.K+r*n.pK))*n.expLambda
-           )/cast_positive(+1u-h*n.L/2u)*(r+1u)*n.pK*pow(h,3u)/4u; }
-template<> FloatDPError twoparam_worstcase_error<SingularInput>(C1Norms const& n, PositiveFloatDPValue const& h, FloatDPError const& r) {
-    return ((r+1u)*n.pK*((n.pH*2u*r+n.H)*(n.K+r*n.pK)+pow(n.L,2u)+(n.L*3u*r+pow(r,2u)*2u*n.pL)*n.pL)*n.expLambda +
-            (n.K+n.pK)/6u*((r+1u)*((n.H*n.pK+n.L*n.pL)*3u +(n.pH*n.K+n.L*n.pL)*4u) +
-                           (n.pH*n.pK+pow(n.pL,2u))*8u*(r*r+1u))
-           )*pow(h,3u)/4u/cast_positive(+1u-h*n.L/2u-h*n.pL*r); }
-
-template<> FloatDPError twoparam_component_error<AffineInputs>(C1Norms const& n, PositiveFloatDPValue const& h, FloatDPError const& r, SizeType j) {
-    return (pow(h,2u)*(pow(r,2u)+1u)*n.pK*n.pLj[j]/2u +
-            pow(h,3u)*(n.K+n.pK)*(r+1u)*((n.Hj[j]*n.pK+n.Lj[j]*n.pL)/8u+(n.pHj[j]*n.K+n.L*n.pLj[j])/6u) +
-            n.pK*(r+1u)*((n.Lj[j]*n.L+r*n.pL*n.Lj[j]+n.Hj[j]*(n.K+r*n.pK))/2u*cast_positive(cast_exact((n.expL*(pow(h*n.L,2u)*3u+4u-h*n.L*5u)+h*n.L-4u)/pow(n.L,3u))) +
-                                                 (n.pLj[j]*n.L+r*n.pL*n.pLj[j]+n.pHj[j]*(n.K+r*n.pK))*r*cast_positive(cast_exact((n.expL*(pow(h*n.L,2u)+2u-h*n.L*2u)-2u)/pow(n.L,3u))))
-           )/cast_positive(1u-h*n.Lj[j]/2u-h*r*n.pLj[j]); }
-template<> FloatDPError twoparam_component_error<AdditiveInputs>(C1Norms const& n, PositiveFloatDPValue const& h, FloatDPError const& r, SizeType j) {
-    return (n.Hj[j]*(n.K+n.pK)/2u +
-            (n.Lj[j]*n.L+n.Hj[j]*(n.K+r*n.pK))*cast_positive(cast_exact((n.expL*(pow(h*n.L,2u)*3u+4u-h*n.L*5u)+h*n.L-4u)/pow(n.L*h,3u)))/2u
-           )*n.pK*pow(h,3u)/4u*(r+1u)/cast_positive(+1u-h*n.Lj[j]/2u); }
-template<> FloatDPError twoparam_component_error<SingularInput>(C1Norms const& n, PositiveFloatDPValue const& h, FloatDPError const& r, SizeType j) {
-    return (pow(h,3u)*(n.K+n.pK)/24u*((r+1u)*((n.Hj[j]*n.pK+n.Lj[j]*n.pL)*3u+(n.pHj[j]*n.K+n.L*n.pLj[j])*4u) +
-                                      (n.pHj[j]*n.pK+n.pL+n.pLj[j])*(pow(r,2u)+1u)*8u) +
-            n.pK*(r+1u)*((n.Lj[j]*n.L+r*n.pL*n.Lj[j]+n.Hj[j]*(n.K+r*n.pK))/2u*cast_positive(cast_exact((n.expL*(pow(h*n.L,2u)*3u+4u-h*n.L*5u)+h*n.L-4u)/pow(n.L,3u))) +
-                         (n.pLj[j]*n.L+r*n.pL*n.pLj[j]+n.pHj[j]*(n.K+r*n.pK))*r*cast_positive(cast_exact((n.expL*(pow(h*n.L,2u)+2u-h*n.L*2u)-2u)/pow(n.L,3u))))
-           )/cast_positive(1u-h*n.Lj[j]/2u-h*r*n.pLj[j]); }
-
-
-template<class A, class R> Vector<FloatDPError> ApproximationErrorProcessor<A,R>::process(C1Norms const& n, PositiveFloatDPValue const& h) const {
-
-    Vector<FloatDPError> result(n.dimension(),worstcase_error<A,R>(n,h));
-
-    if (_enable_componentwise_error) {
-        for (auto j: range(n.dimension()))
-            result[j] = min(result[j],component_error<A,R>(n,h,j));
-    }
-    return result;
-}
-
-template<class A, class R> Vector<ErrorType> ApproximationErrorProcessor<A,R>::process(PositiveFloatDPValue const& h, UpperBoxType const& B) const {
-    C1Norms norms = compute_norms(Ariadne::noise_independent_component(_f,_inputs.size()),Ariadne::input_derivatives(_f,_inputs.size()),_inputs,h,B);
-    ARIADNE_LOG(7,"norms: " << norms << "\n");
-    Set<Nat> input_idx;
-    for (Nat i : range(_f.result_size(),_f.result_size()+_inputs.size())) { input_idx.insert(i); }
-    if (is_additive_in(_f,input_idx))
-        norms.pK=mag(norm(_inputs));
-    return process(norms,h);
-}
-
-InclusionApproximatorHandle
-InclusionApproximatorFactory::create(EffectiveVectorMultivariateFunction const& f, BoxDomainType const& inputs, InputApproximation const& approximation) const {
-    if (approximation.handles(ZeroApproximation())) return InclusionApproximatorHandle(SharedPointer<InclusionApproximatorInterface>(new InclusionApproximatorBase<ZeroApproximation>(f,inputs,_integrator)));
-    if (approximation.handles(ConstantApproximation())) return InclusionApproximatorHandle(SharedPointer<InclusionApproximatorInterface>(new InclusionApproximatorBase<ConstantApproximation>(f,inputs,_integrator)));
-    if (approximation.handles(AffineApproximation())) return InclusionApproximatorHandle(SharedPointer<InclusionApproximatorInterface>(new InclusionApproximatorBase<AffineApproximation>(f,inputs,_integrator)));
-    if (approximation.handles(SinusoidalApproximation())) return InclusionApproximatorHandle(SharedPointer<InclusionApproximatorInterface>(new InclusionApproximatorBase<SinusoidalApproximation>(f,inputs,_integrator)));
-    if (approximation.handles(PiecewiseApproximation())) return InclusionApproximatorHandle(SharedPointer<InclusionApproximatorInterface>(new InclusionApproximatorBase<PiecewiseApproximation>(f,inputs,_integrator)));
-    ARIADNE_FAIL_MSG("Unhandled input approximation " << approximation << "\n");
-}
-
-
-InclusionEvolver::InclusionEvolver(SystemType const& system, SweeperDP const& sweeper, IntegratorInterface const& integrator, Reconditioner const& reconditioner)
+InclusionEvolver::InclusionEvolver(SystemType const& system, SweeperDP const& sweeper, IntegratorInterface const& integrator, ReconditionerHandle const& reconditioner)
     : _system(system)
     , _sweeper(sweeper)
     , _integrator(integrator.clone())
@@ -371,13 +199,9 @@ List<ValidatedVectorMultivariateFunctionModelDP> InclusionEvolver::reach(BoxDoma
     ARIADNE_LOG(2,"System: "<<_system<<"\n");
     ARIADNE_LOG(2,"Initial: "<<initial<<"\n");
 
-    EffectiveVectorMultivariateFunction const& F = _system.function();
-    BoxDomainType const& V = _system.inputs();
-    BoxDomainType const& X0 = initial;
-
     StepSizeType hsug(_configuration->maximum_step_size());
 
-    ValidatedVectorMultivariateFunctionModelDP evolve_function = ValidatedVectorMultivariateTaylorFunctionModelDP::identity(X0,this->_sweeper);
+    ValidatedVectorMultivariateFunctionModelDP evolve_function = ValidatedVectorMultivariateTaylorFunctionModelDP::identity(initial,this->_sweeper);
 
     TimeStepType t;
 
@@ -402,20 +226,20 @@ List<ValidatedVectorMultivariateFunctionModelDP> InclusionEvolver::reach(BoxDoma
 
         UpperBoxType B;
         StepSizeType h;
-        std::tie(h,B)=approximators_to_use.at(0).flow_bounds(F,domx,V,hsug);
+        std::tie(h,B)=approximators_to_use.at(0).flow_bounds(domx,_system.inputs(),hsug);
         ARIADNE_LOG(3,"flow bounds = "<<B<<" (using h = " << h << ")\n");
 
         TimeStepType new_t = lower_bound(t+h);
 
         ValidatedVectorMultivariateFunctionModelDP reach_function;
         ValidatedVectorMultivariateFunctionModelDP best_reach_function, best_evolve_function;
-        InclusionApproximatorHandle best = approximators_to_use.at(0);
+        InclusionIntegratorHandle best = approximators_to_use.at(0);
         FloatDP best_volume = FloatDP::inf(DoublePrecision());
 
         for (auto approximator : approximators_to_use) {
             ARIADNE_LOG(5,"checking "<<approximator<<" approximator\n");
 
-            auto current_reach=approximator.reach(F,V,domx,evolve_function,B,t,h);
+            auto current_reach=approximator.reach(domx,evolve_function,B,t,h);
             auto current_evolve=approximator.evolve(current_reach,new_t);
 
             FloatDP current_volume = volume(current_evolve.range());
@@ -451,261 +275,6 @@ List<ValidatedVectorMultivariateFunctionModelDP> InclusionEvolver::reach(BoxDoma
 
     return result;
 }
-
-template<class A> Bool
-InclusionApproximatorBase<A>::operator==(const InclusionApproximatorInterface& rhs) const {
-    return instance_of<InclusionApproximatorBase<A>>(&rhs);
-}
-
-template<class A> Bool
-InclusionApproximatorBase<A>::operator<(const InclusionApproximatorInterface& rhs) const {
-    return this->index() < rhs.index();
-}
-
-template<class A> ValidatedVectorMultivariateFunctionModelType
-InclusionApproximatorBase<A>::reach(EffectiveVectorMultivariateFunction const& F, BoxDomainType const& inputs, BoxDomainType const& domx, ValidatedVectorMultivariateFunctionModelType const& evolve_function, UpperBoxType const& B, TimeStepType const& t, StepSizeType const& h) const {
-
-    TimeStepType new_t = lower_bound(t+h);
-
-    Interval<TimeStepType> domt(t,new_t);
-
-    auto e=this->compute_errors(h,B);
-    ARIADNE_LOG(6,"approximation errors:"<<e<<"\n");
-    auto doma = this->build_parameter_domain(inputs);
-
-    auto w = this->build_w_functions(domt,doma,F.result_size(),inputs.size());
-    ARIADNE_LOG(6,"w:"<<w<<"\n");
-    auto Fw = build_Fw(F,w);
-    ARIADNE_LOG(6,"Fw:"<<Fw<<"\n");
-    auto phi = this->_integrator->flow_step(Fw,domx,domt,doma,B);
-    add_errors(phi,e);
-
-    return this->build_reach_function(evolve_function, phi, t, new_t);
-}
-
-template<class A> Vector<EffectiveScalarMultivariateFunction> InclusionApproximatorBase<A>::build_secondhalf_piecewise_w_functions(Interval<TimeStepType> const& domt, BoxDomainType const& doma, SizeType n, SizeType m) const {
-    auto zero = EffectiveScalarMultivariateFunction::zero(n+1+2*m);
-    auto one = EffectiveScalarMultivariateFunction::constant(n+1+2*m,1_z);
-
-    auto result = Vector<EffectiveScalarMultivariateFunction>(m);
-    for (auto i : range(m)) {
-        auto Vi = ExactNumber(doma[i].upper());
-        auto p0 = EffectiveScalarMultivariateFunction::coordinate(n+1+2*m,n+1+i);
-        auto p1 = EffectiveScalarMultivariateFunction::coordinate(n+1+2*m,n+1+m+i);
-        result[i] = (definitely (doma[i].upper() == 0.0_exact) ? zero : p0+(one-p0*p0/Vi/Vi)*p1);
-    }
-    return result;
-}
-
-template<class A> ValidatedVectorMultivariateFunctionModelDP InclusionApproximatorBase<A>::build_secondhalf_piecewise_reach_function(
-        ValidatedVectorMultivariateFunctionModelDP const& evolve_function, ValidatedVectorMultivariateFunctionModelDP const& Phi, TimeStepType const& t,
-        TimeStepType const& new_t) const {
-
-    // Evolve function is e(x,a,b) at s; Flow is phi(x,h,b)
-    // Want (x,t,a,b):->phi(e(x,a,b),t-s,b))
-
-    auto swp = dynamic_cast<ValidatedVectorMultivariateTaylorFunctionModelDP const&>(Phi.reference()).properties();
-
-    SizeType n=evolve_function.result_size();
-    SizeType b=Phi.argument_size()-(n+1);
-
-    SizeType a=evolve_function.argument_size()-n-b;
-
-    BoxDomainType X=evolve_function.domain()[range(0,n)];
-    BoxDomainType PA=evolve_function.domain()[range(n,n+a)];
-    BoxDomainType PB=Phi.domain()[range(n+1,n+1+b)];
-
-    auto Tau=IntervalDomainType(t,new_t);
-    BoxDomainType XTP = product(X,Tau,PA,PB);
-    ValidatedVectorMultivariateTaylorFunctionModelDP xf=ValidatedVectorMultivariateTaylorFunctionModelDP::projection(XTP,range(0,n),swp);
-    ValidatedScalarMultivariateTaylorFunctionModelDP tf=ValidatedScalarMultivariateTaylorFunctionModelDP::coordinate(XTP,n,swp);
-    ValidatedVectorMultivariateTaylorFunctionModelDP af=ValidatedVectorMultivariateTaylorFunctionModelDP::projection(XTP,range(n+1,n+1+a),swp);
-    ValidatedVectorMultivariateTaylorFunctionModelDP bf=ValidatedVectorMultivariateTaylorFunctionModelDP::projection(XTP,range(n+1+a,n+1+a+b),swp);
-
-    ValidatedVectorMultivariateTaylorFunctionModelDP ef=compose(evolve_function,join(xf,af,bf));
-
-    return compose(Phi,join(ef,tf,bf));
-}
-
-template<class A> ValidatedVectorMultivariateFunctionModelDP InclusionApproximatorBase<A>::build_reach_function(
-        ValidatedVectorMultivariateFunctionModelDP const& evolve_function, ValidatedVectorMultivariateFunctionModelDP const& Phi, TimeStepType const& t,
-        TimeStepType const& new_t) const {
-
-    // Evolve function is e(x,a) at s; flow is phi(x,h,b)
-    // Want (x,t,a,b):->phi(e(x,a),t-s,b))
-
-    auto swp = dynamic_cast<ValidatedVectorMultivariateTaylorFunctionModelDP const&>(Phi.reference()).properties();
-
-    SizeType n=evolve_function.result_size();
-
-    SizeType a=evolve_function.argument_size()-n;
-    SizeType b=Phi.argument_size()-(n+1);
-
-    BoxDomainType X=evolve_function.domain()[range(0,n)];
-    BoxDomainType PA=evolve_function.domain()[range(n,n+a)];
-    BoxDomainType PB=Phi.domain()[range(n+1,n+1+b)];
-
-    auto Tau=IntervalDomainType(t,new_t);
-    BoxDomainType XTP = product(X,Tau,PA,PB);
-    ValidatedVectorMultivariateTaylorFunctionModelDP xf=ValidatedVectorMultivariateTaylorFunctionModelDP::projection(XTP,range(0,n),swp);
-    ValidatedScalarMultivariateTaylorFunctionModelDP tf=ValidatedScalarMultivariateTaylorFunctionModelDP::coordinate(XTP,n,swp);
-    ValidatedVectorMultivariateTaylorFunctionModelDP af=ValidatedVectorMultivariateTaylorFunctionModelDP::projection(XTP,range(n+1,n+1+a),swp);
-    ValidatedVectorMultivariateTaylorFunctionModelDP bf=ValidatedVectorMultivariateTaylorFunctionModelDP::projection(XTP,range(n+1+a,n+1+a+b),swp);
-
-    ValidatedVectorMultivariateTaylorFunctionModelDP ef=compose(evolve_function,join(xf,af));
-
-    return compose(Phi,join(ef,tf,bf));
-}
-
-template<class A> ValidatedVectorMultivariateFunctionModelDP InclusionApproximatorBase<A>::evolve(ValidatedVectorMultivariateFunctionModelDP const& reach_function, TimeStepType const& t) const {
-    return partial_evaluate(reach_function,reach_function.result_size(),t);
-}
-
-Void add_errors(ValidatedVectorMultivariateFunctionModelDP& phi, Vector<ErrorType> const& e) {
-    assert(phi.result_size()==e.size());
-    ValidatedVectorMultivariateTaylorFunctionModelDP& tphi = dynamic_cast<ValidatedVectorMultivariateTaylorFunctionModelDP&>(phi.reference());
-    for (auto i : range(e.size())) {
-        tphi[i].add_error(e[i]);
-    }
-}
-
-EffectiveVectorMultivariateFunction build_Fw(EffectiveVectorMultivariateFunction const& F, Vector<EffectiveScalarMultivariateFunction> const& w) {
-
-    auto n = F.result_size();
-    auto m = w.size();
-    auto p = w[0].argument_size();
-
-    auto coordinates = EffectiveVectorMultivariateFunction::coordinates(p);
-
-    auto substitution = EffectiveVectorMultivariateFunction::zeros(n+m,p);
-    for (auto i : range(n)) {
-        substitution.set(i,coordinates[i]);
-    }
-    for (auto i : range(m)) {
-        substitution.set(n+i,w[i]);
-    }
-
-    return compose(F,substitution);
-}
-
-
-template<class A> Pair<StepSizeType,UpperBoxType> InclusionApproximatorBase<A>::flow_bounds(EffectiveVectorMultivariateFunction const& f, BoxDomainType const& domx, BoxDomainType const& doma, StepSizeType const& hsug) const {
-    return EulerBounder().compute(f,domx,doma,hsug);
-}
-
-template<class A> BoxDomainType InclusionApproximatorBase<A>::build_parameter_domain(BoxDomainType const& V) const {
-    BoxDomainType result(0u);
-    for (Nat i=0; i<this->_num_params_per_input; ++i)
-        result = product(result,V);
-    return result;
-}
-
-template<> Vector<EffectiveScalarMultivariateFunction> InclusionApproximatorBase<ZeroApproximation>::build_w_functions(Interval<TimeStepType> const& domt, BoxDomainType const& doma, SizeType n, SizeType m) const {
-    auto result = Vector<EffectiveScalarMultivariateFunction>(m);
-    for (auto i : range(0,m))
-        result[i] = EffectiveScalarMultivariateFunction::zero(n+1);
-    return result;
-}
-
-
-template<> Vector<EffectiveScalarMultivariateFunction> InclusionApproximatorBase<ConstantApproximation>::build_w_functions(Interval<TimeStepType> const& domt, BoxDomainType const& doma, SizeType n, SizeType m) const {
-    auto result = Vector<EffectiveScalarMultivariateFunction>(m);
-    for (auto i : range(0,m))
-        result[i] = EffectiveScalarMultivariateFunction::coordinate(n+1+m,n+1+i);
-    return result;
-}
-
-
-template<> Vector<EffectiveScalarMultivariateFunction> InclusionApproximatorBase<AffineApproximation>::build_w_functions(Interval<TimeStepType> const& domt, BoxDomainType const& doma, SizeType n, SizeType m) const {
-    auto zero = EffectiveScalarMultivariateFunction::zero(n+1+2*m);
-    auto one = EffectiveScalarMultivariateFunction::constant(n+1+2*m,1_z);
-    auto three = EffectiveScalarMultivariateFunction::constant(n+1+2*m,3_z);
-    auto t = EffectiveScalarMultivariateFunction::coordinate(n+1+2*m,n);
-    auto tk = EffectiveScalarMultivariateFunction::constant(n+1+2*m,domt.lower());
-    auto hc = EffectiveScalarMultivariateFunction::constant(n+1+2*m,domt.width());
-
-    auto result = Vector<EffectiveScalarMultivariateFunction>(m);
-    for (auto i : range(m)) {
-        auto Vi = ExactNumber(doma[i].upper());
-        auto p0 = EffectiveScalarMultivariateFunction::coordinate(n+1+2*m,n+1+i);
-        auto p1 = EffectiveScalarMultivariateFunction::coordinate(n+1+2*m,n+1+m+i);
-        result[i] = (definitely (doma[i].upper() == 0.0_exact) ? zero : p0+three*(one-p0*p0/Vi/Vi)*p1*(t-tk-hc/2)/hc);
-    }
-    return result;
-}
-
-
-template<> Vector<EffectiveScalarMultivariateFunction> InclusionApproximatorBase<SinusoidalApproximation>::build_w_functions(Interval<TimeStepType> const& domt, BoxDomainType const& doma, SizeType n, SizeType m) const {
-    auto zero = EffectiveScalarMultivariateFunction::zero(n+1+2*m);
-    auto one = EffectiveScalarMultivariateFunction::constant(n+1+2*m,1_z);
-    auto pgamma = EffectiveScalarMultivariateFunction::constant(n+1+2*m,1.1464_dec);
-    auto gamma = EffectiveScalarMultivariateFunction::constant(n+1+2*m,4.162586_dec);
-    auto t = EffectiveScalarMultivariateFunction::coordinate(n+1+2*m,n);
-    auto tk = EffectiveScalarMultivariateFunction::constant(n+1+2*m,domt.lower());
-    auto hc = EffectiveScalarMultivariateFunction::constant(n+1+2*m,domt.width());
-
-    auto result = Vector<EffectiveScalarMultivariateFunction>(m);
-    for (auto i : range(m)) {
-        auto Vi = ExactNumber(doma[i].upper());
-        auto p0 = EffectiveScalarMultivariateFunction::coordinate(n+1+2*m,n+1+i);
-        auto p1 = EffectiveScalarMultivariateFunction::coordinate(n+1+2*m,n+1+m+i);
-        result[i] = (definitely (doma[i].upper() == 0.0_exact) ? zero : p0+(one-p0*p0/Vi/Vi)*pgamma*p1*sin((t-tk-hc/2)*gamma/hc));
-    }
-    return result;
-}
-
-
-template<> Vector<EffectiveScalarMultivariateFunction> InclusionApproximatorBase<PiecewiseApproximation>::build_w_functions(Interval<TimeStepType> const& domt, BoxDomainType const& doma, SizeType n, SizeType m) const {
-    auto zero = EffectiveScalarMultivariateFunction::zero(n+1+2*m);
-    auto one = EffectiveScalarMultivariateFunction::constant(n+1+2*m,1_z);
-
-    auto result = Vector<EffectiveScalarMultivariateFunction>(m);
-    for (auto i : range(m)) {
-        auto Vi = ExactNumber(doma[i].upper());
-        auto p0 = EffectiveScalarMultivariateFunction::coordinate(n+1+2*m,n+1+i);
-        auto p1 = EffectiveScalarMultivariateFunction::coordinate(n+1+2*m,n+1+m+i);
-        result[i] = (definitely (doma[i].upper() == 0.0_exact) ? zero : p0-(one-p0*p0/Vi/Vi)*p1);
-    }
-    return result;
-}
-
-template<> ValidatedVectorMultivariateFunctionModelType
-InclusionApproximatorBase<PiecewiseApproximation>::reach(EffectiveVectorMultivariateFunction const& F, BoxDomainType const& inputs, BoxDomainType const& domx, ValidatedVectorMultivariateFunctionModelType const& evolve_function, UpperBoxType const& B, TimeStepType const& t, StepSizeType const& h) const {
-
-    auto n = F.result_size();
-    auto m = inputs.size();
-
-    TimeStepType intermediate_t = lower_bound(t+hlf(h));
-    TimeStepType new_t = lower_bound(t+h);
-
-    Interval<TimeStepType> domt_first(t,intermediate_t);
-    Interval<TimeStepType> domt_second(intermediate_t,new_t);
-    auto doma = this->build_parameter_domain(inputs);
-
-    auto e=this->compute_errors(h,B);
-    ARIADNE_LOG(6,"approximation errors:"<<e<<"\n");
-
-    auto w_hlf = this->build_w_functions(domt_first,doma,n,m);
-    ARIADNE_LOG(6,"w_hlf:"<<w_hlf<<"\n");
-    auto Fw_hlf = build_Fw(F,w_hlf);
-    ARIADNE_LOG(6,"Fw_hlf:" << Fw_hlf << "\n");
-
-    auto phi_hlf = this->_integrator->flow_step(Fw_hlf,domx,domt_first,doma,B);
-    auto intermediate_reach=this->build_reach_function(evolve_function, phi_hlf, t, intermediate_t);
-    auto intermediate_evolve=this->evolve(intermediate_reach,intermediate_t);
-
-    auto domx_second = cast_exact_box(intermediate_evolve.range());
-
-    auto w = this->build_secondhalf_piecewise_w_functions(domt_second,doma,n,m);
-    ARIADNE_LOG(6,"w:"<<w<<"\n");
-    auto Fw = build_Fw(F, w);
-    ARIADNE_LOG(6,"Fw:"<<Fw<<"\n");
-    auto phi = this->_integrator->flow_step(Fw,domx_second,domt_second,doma,B);
-    add_errors(phi,e);
-
-    return this->build_secondhalf_piecewise_reach_function(intermediate_evolve, phi, intermediate_t, new_t);
-}
-
 
 struct IndexedFloatDPError
 {
@@ -855,7 +424,7 @@ InclusionEvolverConfiguration::InclusionEvolverConfiguration()
     maximum_step_size(1);
     maximum_enclosure_radius(100.0);
     enable_parameter_reduction(true);
-    approximations({ZeroApproximation()});
+    approximations({ZeroApproximation(),ConstantApproximation(),AffineApproximation(),SinusoidalApproximation(),PiecewiseApproximation()});
 }
